Add New GitHub Marketplace Plan to `sync_plans`
Change in pricing requires the creation of a new github marketplace plan. These plan ids are hardcoded into the sync_plans function currently.

This PR is just a quick fix for getting the new plan id in. It is highly recommended to checkout this branch and provide a more scalable solution (e.g., better globals, an env var, etc) for the future worker codebase.