Improving ways of detecting plist
This comes from a case in our own uploaded reports where a plist file came with the pathname as

/home/circleci/repo/services/report/tests/unit/xccoverage.xml

Our logic currently only deals with `xccoverage.plist` files.

We need to also detect plist when the file is not strictly ended with plist.

Also, adding standardized plist handling so it is easier to find where it happens (I wrote it and for a solid 5 minutes I couldnt find where I was using it)