Catching exception for when commit is unreachable
When commit is unreachable, we might still want to try to process the commit. Or at least not crash in the beginning of the processing

So if the commit is unreachable on the git provider, the commit yaml file will also be. Adding an exception handling to deal with that case