Instantiating only one storage client per process
We are trying to make it so we only need to instantiate one StorageService per worker process. The idea is that different clients (like GCP storage SDK) might have initialization steps that are costly and could make us lose some time creating one per each task.

So if we can make it so:

 - Each worker instantiates new StorageClient s only when needed
 - Tests dont get broken because they accidentally use another test storage client

We should be good to go on this.

To be clear, and explain why the change, we saw recently that switching from minio to GCP maade reading files from storage go from 300ms-600ms to 500ms-1200ms per read, possibly because of that extra token fetch that is being done once per client instantiated (check services.archive.read_file.95percentile metrics over Feb 2020 to see)