Showing proper diff coverage on comment
When porting the code from legacy we missed that they did 'report.apply_diff(diff)' outside the comment.py notification code. This apply_diff had the side effect of populating the diff variable with a totals key.

So what we are doing here is calling the 'report.apply_diff' ourselves and using the right variable to determine the content of the line '> The diff coverage is X.'