Adding relative path calculation on pathfixing
We have some situations where the current path of the report file is relevant because the report uses relative paths to describe the covered files.

For now we will only use this for logging purposes. After we collect enought data on how this will disrupt users, we can decide what to actually do