Provide per-file diff information on comments
When we call apply_diff with save=False, the system does not properly adds information to each file dict telling them what diff coverage is

This makes we not properly tell the user information about impacted files. We either miss some of the files that were affected by the pull changes, or not properly calculate the impact of those diffs.

The only change we need to do is the '_save=False', the call already fills that information into the dicts that we already use for impact calculation.