Fixing diff for files that have no coverage
There are some cases where a certain file is present at at Report (ie. has coverage information), and also is at the diff, but the diff information does not overlap with the changes.

When this happens the diff['totals'] dict for that file will have diff['totals'].coverage == None. Which is fine, but when this value is used for sorting later, we have issues that only appear
    on Python3, since on python3 None is not comparable to other values (on Python2 it was so it never was an issue before.

Its also weird that the value was of type str before, which makes the comparisons really weird since they are supposed to be numbers.

Here is the error:

Traceback (most recent call last):
  File "/worker/services/notification/__init__.py", line 93, in notify_individual_notifier
    res = await notifier.notify(comparison)

  File "/worker/services/notification/notifiers/comment/__init__.py", line 85, in notify
    message = await self.build_message(comparison)
  File "/worker/services/notification/notifiers/comment/__init__.py", line 282, in build_message
    return self._create_message(comparison, diff, pull_dict)
  File "/worker/services/notification/notifiers/comment/__init__.py", line 491, in _create_message
    for line in starmap(tree_cell,sorted(files_in_diff, key=lambda a: a[3])[:limit]):
TypeError: '<' not supported between instances of 'str' and 'NoneType'