Working with parent-child race conditions on CF
We have had a couple of customers using CF where they cannot fully control the order of uploading reports. So sometimes parent and child commits might get their first uploaded-reports in unusual orders.

Sometimes it might not even be wrong orders, sometiems they are close enough with enough processing-time variance so the child gets started processing before the parent.

When that happens, its not a big deal, except for the fact that the parent commit is now relevant for the child commit CFing. So three situations might arise:

1) The parent is both uploaded and processed by the time the child commit starts processing. This is fine and normal order of things and the parent will be able to CF the parent sessions (both recently processed and CFed from grand or grand-grand-parents)
2) The child got uploaded and start processing before the parent is even uploaded. That is not great, but all will happen is that the system will use a different parent (possibly grandparent) as the child's parent on the db and for CF purposes. Then the CF will occur as step 1, only with the grandparent as the base
3) The child gets processed after the parent is uploaded but after the parent got its first step of processing. This is the issue we need to solve because at this point the child sees the proper parent, but nothing on the parent report to be CFed, because the parent did not store their results yet.

The solution we are applying will be to keep recursively through parents until we find one that has `commit.state == "processed"` (or until we find no parent commit). That field ensure that the parent has gone through at least one round of processing, and already has at least the flags the customer wants to CF from the previous commits.

Note that this still doesnt solve the issue that the parent might have yet some more report processing to do that would eventually CF to the child, but that case might be an issue, since we might be talking about having to wait for the parent to process dozens of reports, which might take a lot of time and it might be not doable to make the child wait for that (since there is no sure way of knowing the parent is really done).