Removing scope of generic exception catching
We have in our code many places with some code like:

```
try:
    do_this_thing()
except Exception:
    log.exception("Not great")
    # some wrapup code
```

This code arises from the situations where we understand there are bugs on specific parts of that system, but they are either:

- Too corner-case to investigate and deal with
- Already arise from bad user input, so they have low priority
- It's better to crash only that section than the whole code (which has other relevant sections)

One example of such code is report processing. There are many different languages we support, but not all of them are as polished as others. So it makes sense that we let individual processors run inside try...except Exception blocks. If they crash, we can log the crash information and still do the next processor which probably doesnt need to be affected by the exception from this one.

But the end goal is to either remove those or let them live on very specific situations, like the processors/notifiers design pattern.

So we are making two changes in this commit:

1) Removing except Exception in some situations where we already know what the specific exception is
2) Adding some separate clauses to deal with Timeout and other celery exceptions, since we dont want to just swallow those