Trying to pin down cause of somw slowness on pull-fetching logic
There seems to be some cases where notify task is seemingly timing out while inside `fetch_and_update_pull_request_information`.

We dont know if that is indeed the case. But the situation that when a hard timeout is hit, the system kills the worker process and a new one has to be started. This is less than ideal since there is a lot of overhead in this. So we are trying to understand what are the items that are taking too long because there is probably something wrong with that

We are also correcting a small issue where a `db_session.begin_nested()`, that needs to be paired with a closing `db_session.commit()` or a db_session.rollback()` was being called with both

Also changing a time limit to what seems to be the appropriate configuration field. We have no idea why the previous mode worked on legacy. But it doesnt work on new.