Commiting to db more often to release locks more quickly
We have a lot of situations where a deadlock is happening
They seem to happen between different parts of the code. Sometimes between two systems, like new-worker and legacy-web.

We still dont know how to quickly eliminate them, but by doing COMMIT more frequently, we think that we can at least release locks more quickly and avoid some of the deadlocks.

We should watch the errors and see if they decrease after we deploy this