Using commits that are on base branch to do proper updating of commits
In the PullSync task there is a step for the cases where the PR was merged. What we do it:

- We take the commits that are in the PR (and in the db) and update their branch in the database. So essentially all those commits move from the PR to the destination branch of the PR. The exact query is:

```
db_session.query(Commit).filter(
    Commit.repoid == repoid,
    Commit.pullid == pullid,
    Commit.commitid.in_(commits_on_pr),
    ~Commit.merged,
).update(
    {
        Commit.branch: pull_dict["base"]["branch"],
        Commit.updatestamp: datetime.now(),
        Commit.merged: True,
        Commit.deleted: False,
    },
    synchronize_session=False,
)
```

This is the correct approach on the "create merge commit" strategy of merging. But it does not work on the "squash merge" strategy of merging, because in the squash merge none of the original commits actually go to the base branch.

So we have to find a way to deal differently with both strategies.

The idea here is to check if any of the commits actually went to the base branch. And we do that by listing the commits in the base branch right now.

We luckily alreay have a `get_ancestors_tree` method on torngit that does exactly that: lists all commits in a tree sctructure. We just have to flatten it and use the list.

Then in an ideal simple case we could just take that list and make sure only the commits at that list move to the base branch.

But because we can't simply fetch all the commits at that branch to check (we only fetch the first page), it's possible that the base branch moves so fast that by the time we check it part of the PR commits could have been merged via merge-commit, but are already out of the first page of commit listing.

So as long as one of the commits is there, we can assume the other ones also are and they were probably in the next page or something. We don't know any method that sends one of the commits from the PR without sending the other ones (while keeping the exact same SHA of that commit)

This also brings us the the limitation of this logic: if the base branch moves so fast that ALL the commits are out of the first page of listing when we check, we will assume (wrongly) that this was a squash commit. Unfortunately we don't know what information could we fetch to detect it (outside listing commits exhaustively)

There are some issues with this solution. though:

- This means that those squashed commits will stay in their original branch. And right now we dont have ways to delete commits from the branch when the branches are deleted (we catch the webhook, we just don't do that part of the handling). And we don't know what are the consequences of this.