Adding retry possibility for loading carryforward flags
There is one race condition situation where we right now dont have a way to properly carryforward:

If commit B has parent commit A, and B starts processing not only before A starts processing, but before A even had a chance to determine who their parents are.

In that situation B tries to look into A for their report to carryforward, sees that A still has status 'pending' and then tries to go to A's parent. But if A did not determine their parents yet, there is no way for B to find someone to carryforward from.

The difference between this and the other race conditions we solved is that we dealt with the case where B is ahead of A on processing, but at A has at least started their 'UploadTask' (first step of processing). In this situation B is already on 'UploadProcessingTask' and A is in the even shorter window of "Already in the DB but did not finish their 'UploadTask' yet".

Our solution here will be to: if B sees a commit in the 'pending' status and with no parents, then B just decides to retry and wait a bit longer.

It might still not be enough. It might be that, due to some weirdness, A never leaves 'pending' state and never gets its parent determined. But then we really have no way of dealing with it anyway