Replacing way of running async loop
We currently run `run_until_complete` on the base of the tasks

We have had some issues where the some of the async tasks called inside the code seem to keep running after a `SoftTimeLimitExceeded` is raised. This leads to situation where:

1) Code is running where it should be stopped
2) Code keeps trying to access resources that are already closed or being used by other tasks now.

Such an example is when a timeout is hit, the task is stopped, and later something inside the nested-not-killed async functions tries to use a database object. The db_session for that object is already closed, so most operations on such object just raise `DetachedInstanceError`

So we need to find a way to make sure the tasks stop running when the "main" code of a task stops.

Luckily, docs for `asyncio.run` say that's exactly what it does:

```
 asyncio.run(coro, *, debug=False)

    Execute the coroutine coro and return the result.

    This function runs the passed coroutine, taking care of managing the asyncio event loop and finalizing asynchronous generators.
```

So we will try to run it.

The biggest risk here is this way of calling the loop not working well with celery (which run_until_complete already seemingly does not). But local tests show it works ok with it

So this might be good, as it will keep tasks from running when they shouldn't, which will both save resources and avoid conflicts