Upgrading celery
We have recently noticed a sequence of cases where the machines are dying (possibly due to MemoryErrors) and they are not being properly removed and replaced with new machines, making the available worker pool smaller than the allocated 32.

This is probably the reason for why we are experiencing queue backups.

The first step here will be to ensure celery is not the culprit of the dying (some memory leak or so), so we re upgrading it to the latest version.

Later we will try other changes