Fixing case where public repo has no particular bot
In some cases, it's possible the public repository might not currently have a valid 'particular bot' (a bot that is associated with that repo or its owner and has admin-like permissions).

While that does make admin calls impossible, on other types of calls, the admin bot is not that relevant. So we should not stop the calls if the repository does not have an admin-bot (it will stop anyway on another point if the repo is not public and does not have a particular bot).