Increasing delays on retry for LockError
Since there is no way to concurrently process multiple reports from the same commit, we have a redis lock to only allow one report to be processed at a time (per commit).

But this has a consequence: there is a limit on how many reports a single user can upload per commit in a given time period and get fast results. For example, if a user makes 100 uploads immediately for a single commit, and each takes 1 minute to process, then it will take at the minimum 1.5 hours to process them all, at the minimum.

If they all make it timely to the system (basically in a 4 second window, as we explain later), then our system can very likely group them and properly schedule one to run after each other. But there are no guarantees that they will all make it in a perfect time window. So the system might, for example, schedule 99 of them together, and then 1 of then might make it late (this is very hypothetical, actual numbers are very different from this). Then the lone report has to be scheduled alone. The system will just schedule it to run normally, without being able to account for the fact that the other 99 reports are also being processed currently.

So let's stay the lone report gets to a worker for processing. The system will notice it is still processing one of the other 99 reports. And it will reschedule (retry) the lone report to be processed later. When that time comes, it will try to be processed again and also be retried, because another report of the 99 is being processed.

The lone report might catch a lucky break and eventually get in the system just between two report batches being processed. But otherwise, it will have to wait for the 99 minutes (assuming 1 minute per report on this customer) to be able to be processed.

By itself, That is not a problem at all. When the time comes the report will be processed (and it is not like it would be processed before 99 minutes if it were part of the big group anyway). But that means the system has to be able to keep retrying for that window of time of 99 minutes. The system has to be essentially able to retry a lot of times to account that each report batch can take 200 seconds and there might be some 30 of them lined up.

Also, this was a relatively happy case. It might happen that none of the 100 reports are able to be grouped together, and are all scheduled alone. They might all make just inside the right time interval to not be batched with any of the other reports. This sounds rare, but is not that hard, because our "scheduling" tasks are not optimized to group more, but to ensure all uploads are scheduled. There is one "scheduling" task happening per upload, which means there will be 100 "scheduling" tasks for this case. If all reports are uploaded in 5 minutes, for example, each "scheduling" task will on average group the reports uploaded on a single 3-second window. And if all the reports are uploaded in 1 minute, each "scheduling" task will be able to group reports from a single 0.6 second time window. It is not hard to calculate there is just no win here if the uploads are evenly uploaded in those time windows. That's why we usually tell the customer "Upload them all at once (the first 4 seconds after the first upload are always able to be grouped together due to a minimum delay) or upload them farther apart (so by the time an upload happens, the previous ones are already finished)"

So on that case, there will be 100 reports, in no way aligned with each other, all of them essentially retrying at similar times (because retry times are always the same) and conflicting a lot. It will be almost guaranteed that a few of them might stop retrying before getting to get processed.

There are a lot of ways to make this work better:
- Try to measure/guess when the "already scheduled" set of reports is going to finish, and schedule the next reports to be run only after that time
- Do not allow any new report scheduling until that commit has all its "already scheduled" reports finish
- Allow the user to set rules for when we can start processing uploads
- Increase the initial delay from 4 seconds to something much higher

All those solution might have a few consequences, or a big design change. So we will not do them for now.

So for now, we will just tell the system to wait more, retry more and randomize the wait times a bit. LockErrors are already a sign of customers with heavy usage, so it makes sense to assume they will upload a lot more than the expected and their reports might be a lot bigger than the average.

The new retry times are clearly overkill (up to 15 retries, with the 15th retry taking up to 1800 hours), but we will be logging how many retries are being needed so we can later decrease them for more sensible numbers. Doing this overkill has no cost for us, and is better than doing a change for a enterprise customer and later determining it was not high enough.