Removing generic exception catching inside notifier
Notifiers are not supposed to catch general exceptions, specially if they are not aiming to handling them. This job is already done by the 'NotificationService'. By doing that, the notifier does not handle the possible problems, does not give the 'NotificationService' the appropriate information to continue, and does not allow the NotificationService to log the possible problems with the appropriate amount of information.

Therefore we are removing this exception catching.