Trying to make the scheduling task a little less eager
We recently made a change to allow the UploadProcessingTask to not suffer too much about `MaxRetriesExceededError`. That change basically allowed for more retries.

While that worked (we don't have any more of such errors occurring), a sife-effect occurred: clients who were suffering from such problem were mostly heavy users. And they uploaded dozens of reports. Now, with each report being properly given a chance to be processed, the amount of reports retying and renunning increased by a lot. And in some clients, those were reports that were failing already for a lot of reasons.

So instead of picking 100 reports, scheduling them in order, and failing them as soon as one of them fails too many times, we are retrying all the 100 of reports, up to 11 times each (for acquiring a lock), and then retrying a few more times because they seem to consistently suffer from Soft Timeouts.

Basically, we are taking the reports of customers who tend to suffer more from tiemouts and retrying them X times more. (dont know what the value of X is exactly)

So we are trying our first secondary measure mentioned on commit `a767cf97853b7ef85a83fde49d6db094d24d1d7c`:

- - Do not allow any new report scheduling until that commit has all its "already scheduled" reports finish

We are not exactly blocking new reports from being scheduled forever. We are just, for now, making it so if any "already scheduled" report is currently runnig, the scheduler task has to at least wait 200 seconds to schedule a new job.
We don't think this is going to totally solve the issue, but our hope is that this will give some time for `UploadTasks` to allow for more jobs to accumulate before being scheduled, making grouping happen a bit better.