Removing temporary lock
We added a temporary lock because we wanted to move the cerryforward initialization to happen in the UploadTask, and we needed to ensure the UploadTask one and the UploadProcessorTask one did not happen at the same time. We also could not just remove the CFF initialization from the UploadProcessorTask in case a lucky commit got there in the wrong time and did not get CFF initialized at all (old code on UploadTask + new code on UploadProcessorTask would cause that).

So now we can remove it, since the UploadTask is already doing it.