Notificatiosn improvements
Trying some notifications improvements. The core problem is that checks, different from other notifiers, is not atomic. Other notifiers will either have a result missing or fully there. Checks can have a created-but-not-completed moment.

And in one way or another, checks is taking longer. The first impression is that the API calls are taking longer. But it could just be that the logic is taking longer and is somehow masking true API call times.

This is specially aggravated when users have many status checks (like one per flag). 45s works fine for most two-statuses cases. But the report logic is very slow when it comes to calculating per-flag totals and diffs and such. A full refactor of the reports model would be needed to fix this. Because 45s should be enough for a big number of statuses.

As we can't do the right fixes, we can do some small improvements:

1) Increase timeout from 45s to 60s - Might be a slight hit on number of needed pods. But should be fine and could give key customers the enough time window to unblock.

2) Decrease parallelism and only run three notifiers at a time - this could be the above-mentioned masker of true api call times, as many paralell notifiers running CPU at the same time could trick the http lib think the reponse did not come in time. Note that this doesn't help softtimeouts, as the same CPU-heavy logic that happened before will have to happen now. But it will avoid situations where many checks start but never finish.

3) Add some metrics and some logging.

4) Add a lock that could improve some cases where many notifiers call `get_diff` at the same time

None of those things are true solutions. But htey might help mitigate some of the problems.