Trying to look at the commits branches
Trying to look at the commits branches to see if they were alredy merged.

As we talked in the other PR, we didn't find endpoints that gave us all the information we needed for commit branches on Github (also Bitbucket, but we don't have many problems on that yet.

So the idea is to check this 'non-exhaustive' list of branches. If repo-default-branch is on that list, that means that commit is already in the main branch, and we can call it 'merged=True'. This will keep commits that are already uploaded on master (and therefore are never moved there by a PR) to be stuck there and not accidentally move to other branches