Dealing with Github returning 403 on installation token retrieval
We don't know whether that's new, or whether it's just happening more now. But when fetching installation tokens from github, we sometimes get a response:

```
{"message":"This installation has been suspended","documentation_url":"https://docs.github.com/rest/reference/apps#create-an-installation-access-token-for-an-app"}
```

with a 403 status. and it's happening more often, and it is just crashing the code.

We already raise a RepositoryWithoutValidBotError on 404s. Since the semantics are similar (a token cannot be fetched from a user we believed to ha integration installed), it makes sense to just extend that behavior to this too.

Ideally, on those situations, we would set the `using_integration` from the repo to `False`. But that requires a more careful change because that function right now does not expect to have db side-effects. So we are making the smaller change for now and will think about this bigger change later