Splitting assignment
There is this assignment in this code

```
filename = file_name_replacement[filename] = fixed
```

This is a cool feature of python, to do two assignments at once. But what the developer missed was the order in which that happened:

They thought that first the mapping `file_name_replacement[filename] = fixed` would be done (file_name_replacement is basically a cache here), and then `filename = fixed`

But the actual order is the opposite: first it does `filename = fixed`, and then `file_name_replacement[filename] = fixed`. But since `filename` and `fixed` are the same now, the second assignment is essentially `file_name_replacement[filename] = filename`. Which is an identity mapping and is useless to the cache.

Fixing this should not change any behavior (except for some very crazy corner cases), but it might improve performance a bit (although it seems it is not much)