Start using Rust-based report on notifications
Making the NotifyTask (and only the NotifyTask) start using the `ReadOnlyReport`, which is an implementation of Report that uses rust functionality for gaining performance improvements.

For now the class itself still delegates all its responses to an internal `Report`, which is the implementation of `Report` we already use.

But it will also call some internal functions in rust to start comparing performance and correctness