Upgrading torngit to remove filtering
As described on https://github.com/codecov/shared/pull/62 , we are removing the `__enter__` and `__exit__` functions from the report, which will make it not possible to call

```
with report.filter(...):
    ...
```

with it. The reasoning is that the existing filtering was doing some edit logic in place on the report, which was

1) bad for paralellism
2) having to clear caches every time a filter was modified, which is bad for performance

This approach is also cleaner, because the object returned by report.filter() is the actual filter and report can still be used as originally.

So now we are replacing the `with comparison.head.report.filter(**_filters):` calls with `with nullcontext():` calls. `nullcontext` are no-op context, as described in https://docs.python.org/3/library/contextlib.html#contextlib.nullcontext . This will make the changes easier to understand, and in the future we can remove the no-op in a PR that doesn't have as many changes as this one already has