Catching InvalidRequestError
There seems to be an exception called InvalidRequestError, which once happens, keeps showing up on the same container again and again, bothering all future tasks that try to use the same database session.

It's fairly rare that it happens on wrap_up_dbsession, but once it occurs, it's bothersome. We should catch it and try to clean it