Fixing case where there is an EnrichedPull but no pull
In some cases we determined there is a Pull request associated with the commit, but we are unable to reach it for some reason. In those cases enriched_pull is not none, but `enriched_pull.database_pull` is.

We need to deal with those cases. The idea is to deal like there is no pull there