Fixing what NULL a certain query should use
The core idea is that `sqlalchemy`, when faced with `row.json_column= None`, is unsure whether to set the value to the usual SQL NULL or the json 'null' value. So we need to adjust the places that do repo.cache_do_not_use = None and make it use a non-ambiguous SQL NULL

This is done by using

```
repo.cache_do_not_use = null()

```

Unfortunately there is no clean test to assure this works now, since Python will automatically convert both values to `None`. Best we can do is with the exact queries, which I dont really like since they feel very tautological to the implementation. And for those tests we also needed to change `cache_do_not_use` type from JSON to JSONB. `JSONB` is already the type of the object anyway on the database, so this is a good change anyway