Adding `--no-deps` to pip install on second stage
We have a flow on a multi-stage build

- On the build stage we download and "wheel" every package that is going to be needed
- On the runtime stage we just install the generated wheels

This has never been a problem. but now with pip 20.3 new resolver, the more strict rules say that whatever we prepared in the build stage does not exactly match what we told the tool to install on the runtime stage It probably doesn't understand that

```
ribs @ git+ssh://git@github.com/codecov/ribs@v0.1.5#egg=ribs
```

and

```
filename=ribs-0.1.5-cp38-cp38-linux_x86_64.whl
```

are the same thing (despite them having generated the second one from the first one). And then it tries to git clone `git@github.com/codecov/ribs@v0.1.5`, which is not doable since we don't have `git`, and is not wanted since we already have the wheel.

Well, the solution is to just tell pip to only install whatever is in the wheels folder and not try to "guess" what else is needed. This is done by adding `--no-deps`. This answer was found at the bottom of the https://pip.pypa.io/en/stable/reference/pip_install/#hash-checking-mode section