helpers/pathmap: Represent cleaned TOCs as lists of strs.
clean_toc() used to return a CSV row, but that frustrated the ability to
process each path independently. A list of strs gives us exactly the structure
that we want, and it lines up better with later manipulations.

Also, include some extra tests to improve coverage, including the case that I
mentioned in Jira.