Fixing some of the unquote git problems
There are two problems we found when we put the latest `unquote_git_path` code in production:

1) Some files have the format(as seen in the toc):

```
"users/roman/data_quality/bootstrap/Icon\r"
```

This looks like https://apple.stackexchange.com/questions/31867/what-is-icon-r-file-and-how-do-i-delete-them . I see no reason to worry about those files coverage-wise. So we are just ignoring the \r section.

2) Other files have the format:

```
"test/fixture/vcr_cassettes/clickhouse/get_breakdown_values_escaped_\".json"
```

So it's a matter of escaping a '"' inside the filename.

It is unclear to me whether there are other types of chars that will be escaped along the way. There are two possible approaches:

1) We replace `if path[i + 1] == "\\":` with `if path[i + 1] in ("\\", "r", "\""):`. This is basically expanding the exception list of "do not try to convert as an octal".

2) We can also do the opposite and consider that all the unicode octets will start with an "octdigit". So if the sequence after '\' starts with an octdigit we try to do the unicode conversion. Otherwise we append normally as we already do with the "\\" case