Setting updatestamp on pull updates
We found out that we are not appropriately updating `updatestamp` on the Pull object.

It doesn't look like legacy had all its pull updates set the `updatestamp`, but it makes sense that it should have. We don't see any side-effects of setting it.

We chose `datetime.utcnow()` due to the fact that the table uses a timezone-naive column. `functions.now()` (which would call the database now() function) would also work, probably. But it requires an extra round to fetch the new data and it is unclear when it does what (on the test we created for example, it didn't update `updatestamp` the second time)