Using proper default given by the db
Our situation db-wise is not the default one. Our db was designed first, and many python schemas were later created to adapt to it as much as possible.

So we don't want sqlalchemy to try to set nulls whenever it thinks there is not a default.

The idea here is that

```
server_default=FetchedValue()
```

will make it so, if a value was not set on python side, we will send the INSERT with only the set fields, and the db will do whatever defaults they want to do. Then we will fetch that data