Running
```
pip-compile --upgrade
```

which is basically the pip-tools flow for upgrading all packages while still respecting requirement.in version requirements.

So this is mostly newer versions of the same packages as before, but a more recent version.

Notice the requirements.in file, which is the one actively manipulated by us, was not touched on this PR