Removing guards on upload obj not being there
When we first added the upload objects, we added safeguards for the cases where the `UploadTask` ran with the old code (therefore not creating `Upload` objects), and the `UploadProcessorTask` ran with the new code (expecting an `Upload` object).

Those safeguards created Upload objects on the fly.

While this probably was useful during the first hour after we deployed this code, this is not helpful anymore. We checked those logs and the lines `Creating upload on the UploadProcessorTask step` does not show up. Neither do `Commit does not have report row yet` lines

The only risk in removing this are enterprise users that are upgrading directly from below 4.5.5 to this version. Those customers might see some of the jobs fail for a small window of time during the deploy. But since this is a 4.5 -> 4.6 migration, some downtime is expected anyway