Dealing with cases where archive is null
There seems to be some uploads where archive is null. And in our database that column has a not null contraint.

Those are all cases that we are "backfilling" a report in the database, meaning probably  this is a report first uploaded before we started with the new database structure for reports.

It makes sense if those are reports originally using the v2 endpoint and uploading the actual data to redis. A lot of those don't have an actual report in storage and therefore don't have a `session.archive`

It's unfortunate but the only option for those is to actually have an empty string