The joined feature is a super duper weird feature we support which enables users to kind of hide a specific upload-related coverage info from report totals while still keeping it at the line-level
It makes ALMOST zero sense.

And here is its TOTAL usage for the past week:

- The top user made 27 uploads in the past week with flags that have joined=False
- One customer made 15 uploads in the past week with flags that have joined=False
- A third customer made 14 uploads in the past week with flags that have joined=False
- And one customer made 1 upload in the past week with flags that have joined=False

Yes. In the past week 4 customers made uploads using joined=False, with a total of 57 uploads.

I have no idea what the benefit is.

This whole story is to say that there is a small bug related to `joined=False` case. I just wanted to write down this research somewhere.

The point is that the `merge` function only needs to be called with `joined=False` at the actual "merging into master" report. Not the temporary report we hold.

So we are fixing that