Increasing task limit on notify a bit
We currently set notify soft timeout to 60s and hard timeout to 80s.

But for some cases that get soft timeouts the lock living 60s is a problem, because the task gets the exception in just a little above 60s, which means that by the time the lock will be released, it is already not there, and will raise an exception that overrides the timeout one.

This masks the number of timeouts we get.

We already have a fix incoming for the general case. This is just a small band-aid so we can measure some results over the next few days.