Using shared BaseCeleryConfig
We added `BaseCeleryConfig` on shared so worker and api could more easily use the same rules for celery configuration.

Now we need to use this on the worker

We are also moving the

```
logging.config.dictConfig(_config_dict)
```

step from the `setup_logging` to before the initialization happens. This is to make even the `get_config` related logs show up with the proper extra fields on logging tools.

Ideally we should not be doing this in this global place. But we need to do in a place before `initialize_sentry`. So this location needs to be somewhere like a "before everything" signal