Honoring user given base sha
We are implementing here a way for the customer to pick a commit that will itself be the base of a given PR.

This will be done via the API. There, the user will set the `user_given_base_sha`. This will allow the user to set to any field they believe will be usable.

On the worker, we will attempt to use that field. This logic will be done at `_pick_best_base_comparedto_pair`, which will try to choose the best possible base/comparedto pair of SHAs for that pull request.

The new implementation will simply try to use `pull.user_given_base_sha` first in place of `pull_information["base"]["commitid"]`. If that doesn't yield any result, then it tries to do the same with `pull_information["base"]["commitid"]`.

There are a few notes here:

- The implementation will not use the `user_given_base_sha` if it can't find that commit on the git provider
- The implementation will use the user-given value even if that given-value is not exactly on the db. This might mean that `pull_information["base"]["commitid"]` has the exact commit on the db while `pull.user_given_base_sha` only has an approximation on the db