Implementing profiling collection task
This task is meant to search the database for the `ProfilingUpload` that not yet joined into the totals for a given `ProfilingCommit`.

It searches for all `ProfilingUpload` created after the `last_joined_uploads_at`, then processes one by one joining them into the totals result, and then saves the new value for `last_joined_uploads_at`, while storing the totals result into storage.

It then spins a task for the `profiling_summarization_task`, which will summarize that result