Writing normalization task
One of the things we want to do now on profiling is to possibly receive data from different places. This means that the format of the data might vary according to what tool uploaded it.

But our "collection" and "summarization" tasks are still expecting specific formats. So we need a way to convert from the user-inputted formats to the expected ones

In order to do that we are writing a normalization task. The normalization task is meant to be spun on every profiling upload. This way, the normalization can happen in parallel for separate uploads, and the collection task can still do its usual job of collecting multiple uploads and assume they are on the correct format

In addition to do that, we also need now to change the queries to use `normalized_at` instead of `uploaded_at` for `ProfilingUpload`, because we now consider the upload ready to be collected only when it's been already normalized.

There are also a couple of fixes of problems found on the other tasks.