Removing kwargs from some UploadProcessorTask functions
In the past, the way processing was done was:

- UploadHandler gets the user upload params, packs it into a dictionary, appends it to a specific redis key that is a list
- UploadTask goes to that redis key, extracts all elements from the list, and turns it into a list if dicts for the `UploadProcessorTask` `arguments_list` param.
- For each dict `arguments` in that `argument_list`, the function `process_individual_report` was called with `**arguments`

The idea here was that there many arguments, many were useful and many were extraneous. Calling the function with arguments as keyword arguments was a way to let the functions choose what they needed and still be able to ignore the values that were not that useful.

But now all the relevant parameters are present in the `Upload` object, which has its id passed in the same dict. So now all the fields are already saved in the proper db object by the time it reaches the `UploadProcessorTask`.

So it is not needed to use this `**arguments` approach anymore. We are already not using those values but they are still being passed down sometimes.