Testing Rust changes
This is a small call inside `get_changes`. We are wrapping the `get_changes_using_rust` inside a try except block to make sure it doesn't affect anything else.

We are also measuring with metrics to compare the time

For now we are only comparing the paths that are present on each calculation. Later we can actually compare the numbers inside each `Change` object