Codecov Open telemetry usage at the worker
This is a small attempt at using https://github.com/codecov/opentelem-python for open telemetry

It is installing the package using a few envvars:

- `OPENTELEMETRY_ENDPOINT`: The endpoint that the client is going to talk to
- `OPENTELEMETRY_TOKEN`: The authentication token to send the profiling data. Can be create via admin UI for now.
- `OPENTELEMETRY_CODECOV_RATE`: The rate of the samples that will have its coverage tracked. We expect it to be around <1% for more performance-conscious users.

We are also using a regex: `re.compile("run/.*")`. The purpose of this regex is to keep from trying to track tasks on celery on other moments like `apply_async` (which from my perspective is not really worth tracking). We can figure other ways of picking which samples to track, but this seems sensible for now.