Saving notification result only on `notification_attempted`
We currently save almost all notification results. As long as there is a `individual_result["result"]`, a new row is created or updated on the database.

But I believe we should not be doing that when `notification_attempted is False` happens. The reasons are:

1. It is causing a semantic weirdness on the `state` column of the `commit_notifications` table: cases with `notification_attempted is False` also have `notification_successful is None`, which makes the notification be saved with a `pending` state, which is incorrect. There is nothing pending and nothing indicating the notification would be ever done. We would need a proper `not_attempted` state to represent those.
2. Since there is a uniqueness on `(commit_id, notification_type)` tuple, multiple notifications of the same type (ie, "comment" type) overwrites one another. Then in a situation where a PR is created and later closed, we might have a notification comment that is sent (the state would be "success" on this one), and later a notification where no comment is sent (the state would be "pending"). The later one would overwrite the first, and the fact that a notification ever occurred might be hidden by this. It would be better if just the original "success" notification with the proper older timestamp was there.

Both of the situations would point to some level of re-design of the table and feature. But since the current use of the table is only for "User Activation" tasks, those changes wouldn't produce much benefit.

3. Not storing all the "not attempted" notifications better represent of users expect.
  - For example, a particular user has 160+ notifications setup. But in practicality they only expect 5 of them to actually go through on a given commit, because of their particular setup. So it makes sense if (in a hypothetical UI) they would not see those 155+ extra notifications.
  - The user case makes even less sense on this table design, since 150+ of their statuses are of the same type, and therefore they always have only one row for that type anyway (meaning the final stored result depends on which of the 150 project statuses was run last)

All in all, this is being done because I believe that saving all those results take a small unnecessary timing toll.

And since the only feature using it is the above described "User Activation", which only looks at one notification per commit, we don't really need to be saving one extra row for actually unnattempted notifications (no point in retrying something that was not attempted)

All in all, I believe we should be re-doing this logic. This is just a temporary thing to (hopefully) unblock the notifications from the particular 160+ notifications customer.

Also adding some metrics to measure the timing of a particular function