Using get_uploaded_flags
There is a particular piece in our code where we determine whether the flags that were *uploaded* on a given report match the flags that are present on a given status: https://docs.codecov.com/docs/commit-status#flag_coverage_not_uploaded_behavior

The logic itself is fine, but it unnecessarily loops through the same report multiple times. So we implemented this function `get_uploaded_flags`, which caches its results (since it doesn't change at notify time), and are just calling the function many times.

We hope this will save some time, because some customers have 650+ different statuses, and the report has 650+ flags, and very few of them are actually uploaded.

We are also reorganizing the function a bit