Not calling `get_filtered_comparison`
It looks like `get_filtered_comparison` is an expensive operation, because it iterates through all the files in the report. And on certain customers (with 12+ files in the report), this could take a lot of time.

So we are just moving the `get_filtered_comparison` to inside the if so it doesn't get called on the exclude case

As time goes by we can change the `get_filtered_comparison` to not needlessly iterate through all files