Adding proper JSON arrays processing
We found that there is a few places where a customer might upload a json that is:

1. An array
2. Not a format supported by us

And since most of our json processors expect a dict, almost all of them crash while calculating whether the json matches the type of content they parse.

This changes changes the processors themselves to not crash under an array.

Ideally we would also have some try except at the `process_report` function. But for now this alleviates the problem.