Changing unexpected changes in case of new files
When writing the `get_changes` function (the function responsible for finding unexpected changes), the developer made a particular choice.

In the context where the function gets two reports, `head_report_file` and `base_report_file` and a file-level diff `diff`, they added those lines:

```
        if not base_report_file:
            new_files.add(filename)
            continue
```

which meant that, if the `base_report_file` was empty (or None), we should not compare the two files and instead just add the given filename to the `new_files` list (which would always be added to the list of unexpected changes).

This meant that, if a file only has a head report but not a base report, the system just considered this an unexpected changes regardless of the context.

But there are at least two contexts in which this^ is false:

- The file is brand new (tagged on the diff as `diff.get("type") == "new"`). So it's natural that the coverage on new line is brand new, therefore not unexpected. it will be part of the diff coverage, not unexpected coverage.
- The file is not brand new, but it previously only contained non executable lines (maybe it was a file with mostly docstrings). The lack of executable lines make the report for that file be empty (or not there). And then the new change included some executable lines. On this case, all the executable lines that are part of the new report and part of the diff. So all those lines are part of the diff changes, not the unexpected changes.

In both cases, the idea is that if all the new lines are contained inside the git diff, then all those lines (and therefore the file as a whole) are absolutely "not-not-unexpected" changes.

This change addresses those cases.

It's important to notice that this will have basically no impact on the pull request "unexpected files changes" comment section. Since this case happens on files that are present on the git diff, they already are being part of the table due to being in the "files in diff" group of files