Changing "Archived report" logging
When we log `"Archived report"`, we currently log all the sessions that were saved.

While this works ok most of the time, for some customers we get 100+ keys in that dict. This makes it prohibitevely expensive to fully log those cases. UIs like the datadog one and ELK one have trouble even showing it well