Making contextual lines still be marked as unexpected changes
This is related to https://github.com/codecov/shared/pull/239

On git, there are diffs like:

```
         sessions: &Option<Vec<i32>>,
     ) -> diff::FileDiffAnalysis {
-        let (_, lines_on_head) = diff::get_exclusions_from_diff(Some(&diff_data.2));
+        let ((_, lines_on_head), (_, _)) = diff::get_exclusions_from_diff(Some(&diff_data.2));
         let mut involved_lines: Vec<(i32, line::ReportLine)> = Vec::new();
         for line_number in lines_on_head.iter() {
```

Where the diff shows the added lines, the removed lines and some contextual lines.

Some git providers also does this at its API: they show three lines before and three lines after the actual changed lines. Bitbucket has none. Gitlab has an unknown value.

On the python code, it was decided that if there is some unexpected change on those contextual lines, they are not counted as unexpected changes. They are not counted as diff changes either, since they are not part of the actual diff.

On one hand, that makes kinda of sense: calling a change "unexpected" is a high bar, and it makes sense that the lines close to a diff might change coverage because of the diff. So changes in that lines are not quite "unexpected". It also provides a little buffer of safety: "if you are too close to the diff, don't worry that we won't call this change unexpected"

On the other hand, there are a couple of problems:
- Without this logic, every change will be properly categorized: it is either "diff-related change" or "unexpected change". One could technically see the total file coverage drop of X%, and cleanly say "diff changes was responsible for Y% and unexpected changes were responsible for (X - Y)%". With this logic there is a group of changes that belong to neither, and make the previous affirmation inconsistent
- One could argue that "but it makes sense that this line are not really unexpected". But the "diff" / "unexpected" changes division is pretty clear cut. By saying "it makes sense that this lines would have its coverage affected because of the diff even though they are outside of the diff", there are many implementations that respect this rule, like saying "every line after the first diff could technically be in this 'kinda-expected group'", or "when a return is removed or added, everything after it inside the same function can be considered kinda expected"
- It makes it so the same change on different repos provides different results: on Github there will be more contextual lines and on Bitbucket there will be less. Even on local `git diff`, users can pass how many contextual lines they want ( https://stackoverflow.com/questions/13627598/how-to-get-git-diff-with-full-context ), so it is pretty estabilished that contextual lines are not objectively defined.

So there are a few reasons to change it to not consider contextual lines