Upgrading black to stable
```
➜  worker git:(master) ✗ black --version
black, 22.1.0 (compiled: yes)
```

The command we ran was actually:

```
black --skip-magic-trailing-comma .
```

The skip comma option has a slight reason to be.

In the past, whenever a line with multiple parameters had its length go past a certain threshold, black would automatically break the line and add commas after all parameters including the last one. This was done in order to make future parameters additions to not need to touch the previously last line unnecessarily (which makes the git diff harder to read), because the comma was already there.

And later, we would modify the code for it to have lower line size again for some reason, and the comma was not removed. This was fine because the comma is meaningless.

But now the comma in the end means "force-split this line". The `--skip-magic-trailing-comma` option not only skips that option, but also removes the unnecessary comma and line breaks. this will make future checks not need the option.