Implementing ReportBuilder
ReportBuilder is a structure that will make handling of reports more uniform. It will allow for example to pass labels into all lines of an individual report.

The general goal is to change the signature of processors from

```
    def process(
        self, name, content: Element, path_fixer, ignored_lines, sessionid, repo_yaml
    ) -> Report:
```

to

```
    def process(self, name, content, report_builder):
```

as the report_builder will contain the other variables

One of the important elements here is SpecialLabelsEnum. They are a class of labels that are special, and will need special treatments (be converted into the proper labels) once the final report is outputted