Implementing PyCoverage
This the implementation of the PyCoverage json output command( https://coverage.readthedocs.io/en/6.4.4/cmd.html#json-reporting-coverage-json )

The big advantage of this format is that it already includes the test names that executed each line. That makes it "natively" support labels (test names)

This processor is also going to be hopefully a mould of how to implement the other processors using the `ReportBuilder` class.

Notice the `return report_builder_session.output_report()` right at the end. Once we have all processors doing that, we can extract this from the class and have the `process` return a `ReportBuilderSession` object always (which is going to be more flexible than the legacy `Report` object)