Implementing Git Diff Parser
This is a helper function that will make it easier to understand diff and operate on them

We currently have this dict structured described by `diff_to_json` logic on torngit. It produces a dict that, while properly describes the whole change, it still requires logic on top of it to properly analyze what changes happened on the diff

So we are creating DiffChange, which is going to be a class that holds all the relevant data and has some small functions to help process the diff better. Each diff can be represented as an array of `DiffChange`s.

It will also have `map_base_line_to_head_line`, which can map which line is the corresponding line to a given one on base (or head).