Creating empty label analysis Request task
This is the task that will process the "iterative testing" results.

While we are still finalizing the other parts, I want to put the empty task in here so we can have a skeleton on how this will work. There will be a few changes down the line, but this should be the core of how this is calculated