Changing RSpecProcessor to SimplecovProcessor. (#911)
This change is expected to bring more clarity to what processors expect.
We have 2 processors that can handle RSpec data: VOneProcessor (the codecov native format) and SimplecovProcessor (formerly RSpecProcessor). Apparently the V1 format is convinently very close to RSpec so that it could handle it as well.

The former RSpecProcessor actually expected json files formatted by simplecov-json, a formatter plugin for Simplecov, the code coverage library that (apparently) most people use to get coverage info out of Ruby. That caused a confusion as to why are there 2 processors that handle RSpec? So by renaming RSpec to Simplecov we hope to make it extra clear that Simplecov is to handle JSON files formatted by Simplecov and V1 those formatted by god-knows-who cause I couldn't find it :3