Make sure different notifiers stay consistent with config (#954)
The `StatusNotifier` had an extra check in `can_we_set_this_status`. That would restrict branches for which the status is delivered in the config if a list of branches was specified. Apparently `ChecksNotifier` didn't use that behavior, which would potentially cause different status showing up in the status and the checks. To maintain consistency we're adding the `can_we_set_this_status` check to `ChecksNotifier`.