Schedule health check task to run every 10s in its own queue. (#952)
* Schedule health check task to run every 10s in its own queue.

Related ticket: [CODE-2100](https://codecovio.atlassian.net/browse/CODE-2100)
Definitions are in `worker` rather than `shared` just because all CRON task definitions are in worker.
The dedicated queue is hardcoded, as it states in the ticket. The objective is to make sure health checks are not stuck in a queue that's full.

* force healthcheck queue to run for all workers.

This will ensure that the healthcheck is executed and that the healthcheck queue doesn't fill up without any workers to consume from it. The healthcheck task itself is pretty lightweight that this shouldn't be a burden for other tasks running on that worker