Add metrics.timer to `build_payload` function (#988)
* Add metrics.timer to `build_payload` function

Related ticket: [CODE-2304](https://codecovio.atlassian.net/browse/CODE-2304)

Investigating recent issues that a self-hosted customer is having we felt a lack of monitoring data in the build_payload process of checks notifiers. Adding metrics.timer decorator to the build_payload function will help us get more visibility on exactly what part of the notification process might be taking longer than expected.

* Move metrics.timer inside async function

There's a reasonable chance that doing `@metrics.timer()` on a `async` function will return only the time to generate the `Coroutine` that is the async function and not the time taken to actually run it (which happens when you `await` the async function).

So to be sure that we are timing the right thing the best is to put the timer inside the async part of the function.