Use commit yaml when tagging manual critical files (#1006)
Related ticket: [CODE-2296](https://codecovio.atlassian.net/browse/CODE-2296)

Recentrly we created manual critical files, where a user can specify paths in their yaml, and files matching those paths will be marked as critical in the PR comment. The issue we had is that we were using the repo yaml, not the commit one. That means that only paths in the master branch were being considered, causing confusion to users. It also means they need to merge paths without knowing if they're correct or not.

To resolve these issues we will be using commit yaml now. That has the added issue of getting the commit yaml in the overlay, which requires an extra call to the provider.

The funny thing is that we already do that in the notify task - which is the task that calls the PR comments anyway. So we can reuse that yaml by passing it down all the way to the overlay. That means adding it to the comparison though. I'll argue that it's not so far outside the purpose of a `Comparison` object to have the yaml of the head commit in it, so the change is there, but can be discussed.

The other issue is that getting the yaml is an async operation and so far the overlays were all sync. So there are changes to accomodate that as well.