Implement `removals_only` behavior for `removed_code_behavior` (#977)
* Implement `removals_only` behavior for `removed_code_behavior`

Related ticket: [CODE-2162](https://codecovio.atlassian.net/browse/CODE-2162)

Removed code behavior is the ability to get a passing status on `project` according to some criteria.
The 1st method is in code removal, i.e. if a diff consists only of code removal, we can pass the project status.
No coverage addition and no unexpected line changes are permitted.

There was some refactoring to `get_project_status` to accomodate the fact that statuses can be "saved" by the removed_code_behavior. Return is only at the end of the function now, having the `if/else` structure populating the `state` and `message` vars. I felt like it doesn't make sense to use removed_code_behavior when there's no base report to compare, or no coverage on base report, so no change there. Finally, there was a piece of code in the actual comparison that would write the coverage message with "... remains the same compared to ..." that was actually being overwritten, so I just removed it.

The second big change is that `get_project_status` is now an `async` function. That was a necessary change because of `comparison.get_impacted_files()` which is async as well.

* Wrap original `get_project_status` function, only calculate removed_code_behavior if status is failing, not all files needed removed code in 'removals_only'.

To preserve the original calculation of the project status we're wrapping that (sync) function in an (async) function that will get the project status as previously calculated (only change is the function name) and possibly apply the removed_code behavior configured IFF the status is failing.

Also for the `removals_only` behavior not all files in the diff need necessarily to have code removed, but **all** of them need to have no code addition nor unexpected changes.