Revert "Revert (#1017) - upload finisher task is failing the dynamic routing (#1018)" (#1019)
This reverts commit 28f8791442d5c9d9f5df7add414c1decf5e944ed.
And re-enables the dynamic task router.... again.

This time the UploadFinisherTask, when calling ComputeComparison task was passing the arguments via `args` instead of `kwargs` as everywhere else. This breaks the expectations of the task router (that depends on `kwargs` to find the user plan).

Since the trend all over the worker is to pass arguments via `kwargs` I don't think it's a problem if we just fix this one place. If needed we can look into making the task router more robust, though.