Removing special processing of "global lines"
We currently have special processing where, when a "global line" is found, the system automatically replaces it with every other test in the the same processing

While that is reasonable behavior to have, it is increasing a lot the size of the resulting chunks file. There is no reason why we can't move this processing to on the fly the label processing task (or even use this extra "this line is global" information to our benefit)

This will remove the extra size from tasks that dont need it

This will make the label analysis slightly wrong on "global lines" cases until we fix it. But we should be doing that soon.