Splitting different tests into different datapoints
We accidentally made all the tests into the same datapoint. That is not part of the design of this structure. While that could be part of the design, we would have to change a few things. So we are changing this for now

To accomodate things properly, we are also changing `labels` to `labels_list_of_lists`, so the system knows to create separate datapoints when it sees a list of list of labels