Adding component to YAML
Adding components to this repository

The idea is to create components so we can dogfood the feature a little.
It's also going to be helpful when we do some iterative testing changes

The components we are creating are two:

1. Non-Test Code: Tracking coverage outside of test files only. We already do that in general, but due to Iterative Testing, we will need to include test file coverage in the general coverage. So I want to keep at least one place showing coverage excluding those files.
    - As a side effect, once we stop ignoring the test files from the global level, we will also be able to tell if the filters work since the coverage on this component should be the same after the change
2. Code outside of tasks: Code outside of task folder should be well tested by unit tests (tasks are ok to be mostly tested in integration tests as some things might be harder to test in unit)

I am also creating statuses on both of those cases. Hopefully we can see if that's useful or not