Introduce announcement section for gh app login in pr comment (#1026)
* Introduce announcement section for gh app login in pr comment

Related ticket: [CODE-2442](https://codecovio.atlassian.net/browse/CODE-2442)

We will soon be changing login to use out GitHub app, not OAuth. This change is important, and when it happens users that don't have GH App installed in their orgs will start to have issues with codecov (permissioning of tokens and whatnot)
To increase awareness to this change we are adding an announcement to github users that currently don't use the integration.
Said announcement goes right after the header section to really bring attention to it. It will be annoying to some, but we need the information to get to them.
enterprise customers will NOT receive the annoucement cause their login schema is different and they do whatever they want.

Most changes in this commit have to do with changing tests, though.
Unit tests didn't use the integration (still don't), but instead of changing all asserts for something that's likely temporary I just faked an integration for them. That's more closely to what we want in the future anyway so I think it's fine.
For integration tests I just left as is cause I needed to test my own changes anyway, so we know the annoucement work, yay.

* Include 'Learn More' link; don't use AnnouncementSectionWriter