shift lines of carryforward reports when creating them (#1050)
* shift lines of carryforward reports when creating them

Related ticket: [CODE-2541](https://codecovio.atlassian.net/browse/CODE-2541)

We have the ability to shift the lines of a report from a diff extracted from the git provider. So far we haven't been doing that because when something is changed the user most likely will run all tests in that file and over-write any line shifts we might have done.

This situation changes with Smart Test Selection because we might not run all tests for a file, and so do a partial carry-forward. Then it will be important to shift the coverage info carried forward so it matches the files in the new commit.

In order to do that we get the diff from the parent-commit (where the coverage info being carried forward comes from) and the current commit (the commit we're creating a report for) and just shift them.

This might reduce the number of unexpected changes as well, although this is more wishful thinking than anything else.

* Improve exception logging for shift_lines

Adding test and improve exception logging for shift_lines_by_diff when reating carryforward reports.
Importantly, also re-raising the SoftTimeLimitExceeded in the unlikely but not impossible event that we get one exactly here.
Because there's a generic `Exception` bucket there this is necessary.

We don't want to continue processing if we receive that exception.

* Elegantly re-raise SoftTimeLimitExceeded + include test from `create_new_report_for_commit` that shifts carry forward from parent commit