Fix `_get_table_data_for_components` when there's no base report. (#1111)
Ticket: [CODE-2705](https://codecovio.atlassian.net/browse/CODE-2705)

The gathering of data info for components fails with `AttributeError: 'NoneType' object has no attribute 'totals'` for the case when we don't have a base report. These changes address that error by setting the "before" info to `None` if there's no base report.

Unit tests for this case are also included.

I think it's not possible to have the case where the head report is missing, because there would be nothing to report on.