Make pushing to staging branch push to staging env (#1118)
Currently the only way to deploy a new worker to staging is via the master branch (i.e. commits merged in master go to staging)
These changes are to allow that pushing to the `staging` branch also pushes changes to the staging env so we can manually test things without them being merged to master first.