Upload.storage_path can't be null (#1201)
* Upload.storage_path can't be null

Recent changes effectivelly make the storage path of the uploader as null.
Logs from prod show this is not a good idea.
Object `id_ mdH55IgBSxHJFx7UvKva` from ELK says:
```
IntegrityError('(psycopg2.errors.NotNullViolation) null value in column "storage_path" of relation "reports_upload" violates not-null constraint
```

So there were 2 options: drop the not-null contraint or change the null for something.
I think that changing it to be `deleted` is better, because it's a very explicit indication that
there was something there, and it was deleted for a reason.

Also we try to re-write the raw_report _after_ potentially deleting it in the same path.
So now we checked that we didn't just deleted it right before trying to rewrite it, thus prioritizing the delete.
ps.: The reason this was never an issue is because we were assigning `archive_url = None` before,
so it was apparently not changing the upload object. Now it's `upload.storage_path = "deleted"`. We can go back, if preferred.
Note to self: don't trust unit tests all that much

* Revert deleted archive situation to what it was before.
We leave the `upload_storage_path` untouched, as a dangling pointer to nothing.
Only try to re-write readable if not deleted.