Refresh labels before calculating final result (#1210)
This change simply adds a step to try and refresh the requested labels (if necessary)
before calculating the final result in label_analysis.

This comes after the possibility of adding the requested labels AFTER the initial request.
So while we are running much of the label_analysis request, the CLI is collecting the labels and PATCHing them later on.
With the present changes we try to get them one more time before calculating the result.

I know what you're thinking "omg, but what if it takes longer to calculate the labels? Then we won't have them to calculate the final result and all is doomed"
Worry not, for the CLI doesn't (yet) patch the labels afterwards.
But even when it does, the requested labels serve only to calculate the `absent_labels` list. And we can make the same calculation directly in the CLI as a fallback.
So even if we don't have the requested labels all is not lost. The CLI will only have to work a little more.