When creating ReportDetails set `_files_array` (#1216)
Since the move of `_files_array` to GCS (not yet in prod) the name of the field is different.
When testing locally I have issues with creating the `ReportDetails` by setting `files_array` instead of `_files_array`.

The exact error was `AttributeError: NoneType has no attribute commit` while trying to access the files_array.
Apparently `self.report` was `None`. This doesn't make much sense to me, honestly, but it's what I saw.
I suspected that it was because it was the first commit of a repo (actually the first ever) so maybe that was the cause of that strange behavior.

In any case if we set `_files_array` it works, and it doesn't matter much because it will get moved to GCS when we actually set it eventaully.