Fix issues writing files_array to GCS
The immediate error that we saw was that `ReportDetails.report` was `None`. I believe that's because we were not passing the `current_report_row` when creating the `ReportDetails` (see `services/report/__init__.py` changes). We had a similar issue a couple lines down before. That case slipped by.

Then because my env does save reports to GCS I noticed differences between saving data in DB vs GCS. The difference is that the constituent parts of `files_array` in DB returned the classes they belong to (idk how), vs the encoded data that comes back from GCS (json format). In theory that is not an issue, because both formats are pretty much interchangeable, but the less differnces the better. SO I introduced a rehidrate function that wraps the `files_array` in the classes they are supposed to be.

And finally to avoid further issues we will default to writing data to the DB if we can't access the repository from the `ReportDetails` object.