Standarize return of files_array
Using the hydration function we can better standarize the data that is returned when accessing `ReportDetails.files_array`.
The difference was mostly within a Session when you set `files_array` in the DB you still use the data you had before.
Differently, when setting it to storage you always get back the encoded data, because it goes to storage immediatelly.

By standarizing the return we make sure to always use the appropriate inner classes, regardless of where the data comes from.