Add `LabelAnalysisProcessingErrors` (#22)
Currently it's hard for customers of ATS to self support when we are missing data.
This usually leads to the customer seeing all labels as absent, even though their setup is apparently commit. It's usually because we are missing data.

By including error messages (using `LabelAnalysisProcessingErrors`) we can give them more information about what is wrong, so they can better self support.