Move Commit.report to GCS
Moving Commit.report to GCS. The field name is actually 'report_json', but that's not super important.
Adjusting tests as well.

On some occasions I had to change `.report_json is None` to `._report_json is None and ._report_json_files_array is None` (or a logically equivalent expression for the given use case). This is because the default value is {}, so accessing `.report_json` will never return None.