chore: debug files array missing meta
We've been having files_array data missing the 'meta' key.
This started happening after moving data to GCS, so naturally we assume something is wrong.
To help us identify the issue (because most the time you can see the 'meta' key in the data)
we're adding a temporary debug log when saving it.

Part of codecov.platform-team#119