chore: change logic to write data to storage
After moving `ReportDetails.files_array` and `Commit.report` to archive successfilly for Codecov
we want to do the same for some selected customers (in `repo_ids`) and then we will want
to start the process for new fields. Only for codecov.

The way things were before, if you have a field set to write to GCS it is valid for all codecov
repos AND all repos under `repo_ids`, while we would only want that for codecov repos.

So by changing the `master_write_switch` to strings we can have more power over each individual
field. Now we can use:
*  `True` or `codecov_access` to write *only* for codecov repos
*  `restricted_access` to write to all repos in `repo_ids` (+ codecov)
*  `general_access` to write to all repos

Because currently we're only writing for codecov repos the change will not affect any customers

Part of codecov/platform-team#100