chore: Standarize label-analysis error return
Label analysis was returning `None` instead of empty list when some processing error happens.
The CLI always expects lists, so that is confusing for it and can lead to errors.
Instead of adding type checks in the CLI I think it's better to standarize the return values of
label analysis to always be lists. These changes do that.

Part of codecov/platform-team#104