chore: remove requirement of beta groups in ats (#52)
Remove the requirement to configure `beta_groups: "labels"`
in the `codecov.yml` when using ATS.

Moving to open beta we think it's an extra step in config
that will confuse customers that forget to do it. And it's
very easy to miss. So we will remove it.

I change it instead to return True if the user has any flag
configured to use the labels carryforward. This is because
the labels thing changes the report (making it bigger), so
no reason to add that to customers that don't use labels.

codecov/platform-team#104