fix: update archive field cache
* Set varialbe cached_value property name, for the case were a single model might have
multiple archived fields
* Concentrate all archive getting/setting in the `__get__` and `__set__` functions

The test changed happens because now we update the cache on the write. Because of that we
are not doing the encode/decode/rehydrate operations. So the data you put in is the data you get.
On one hand we can do such operations to guarantee consistency. On the other this is no different than
what we used before `ArchiveField` AND such operations might be expensive.