Use get_branch in get_behind_by instead of get_branches
This commit changes get_behind_by to use get_branch instead
of get_branches because of a GH API rate limiting issue.

Instead of listing all branches in a repo via the API we now
fetch only the one we are looking for, in this case the base
branch of an open PR.

This commit also makes changes to the requirements by updating
shared to the latest version in order to use get_branch.

Changes had to be made to existing integration tests as they
we re-run, cassettes were updated.

Signed-off-by: joseph-sentry <joseph.sawaya@sentry.io>