Add templates using jinja2
This commit creates the TemplateService that uses
jinja to get and render templates.

This TemplateService provides the get_template method
to get a rendered template. It takes the file name of the
template as an arg and the template variables as the kwargs.

Templates should be added into the templates directory
at the top level of the repository. This commit creates two
example templates.

Signed-off-by: joseph-sentry <joseph.sawaya@sentry.io>