Create Email class and create SMTP service
Create Email wrapper class over EmailMessage.

Create an SMTP service in order to connect to
an SMTP to send emails.

This service requires services.smtp to be populated
in the install config. For more info: https://github.com/codecov/shared/pull/39

This service provides the open method to create a connection to
the SMTP server.

The send method allows for sending an email specified by an
Email object.

Creates a mock_smtp fixture for testing.

Signed-off-by: joseph-sentry <joseph.sawaya@sentry.io>