chore: don't save errors in `LabelAnalysisRequest` object
In the case where we have some missing information we were saving the errors as part of the result dict (in the database).
There's a dedicated table for the larq errors, they should not be saved in the database.
THese changes fix that adding the errors to the dict to return only after savign the real results.