chore: add warning if only special labels in report
After merging #113 I decided to investigate a little more why we would receive nothing but SpecialLabels in a report.
Turns out that if you generate the reports without `--cov-context=test` but do `coverage json --show-contexts` it does that.

So having a report with only special labels might indicate that ATS is not properly configured,
that is, the tests are not running as they should.

Ideally we would have a better place to surface this to customers, but right now we don't.
So the least we can do is add a warning. In this case, if we look at all the logs in the task,
we will see this detail.