Add templates using jinja2 (#107)
* Add templates using jinja2

This commit creates the TemplateService that uses
jinja to get templates.

This TemplateService provides the get_template method
to get an unrendered template. It takes the file name of the
template as an arg.

Templates should be added into the templates directory
at the top level of the repository. This commit creates two
example templates.

Signed-off-by: joseph-sentry <joseph.sawaya@sentry.io>