fix: better handle missing static analysis info
Fix codecov/engineering-team#650

Recently we made changes to static analysis process of the CLI so it doesn't fails completely
if there are issues analysing a file or uploading it to GCS.

I think this inadvertedly broke the expectation of label analysis that all files in the static analysis
are uploaded and ready to be used. Or some other misterious bug is causing some snapshots to be missing

This is causing issues that the label analysis is not processing correctly. So we are now guarding
against missing snapshots in HEAD or BASE.

Sadly for ats, it only as good as the data it has. And being conservative (and safe) is important to us.
So if we can't make an educated guess about a change, the procedure so far is to just run everything.
That's the fallback, assume the entire file is compromised and needs to be tested again

closes codecov/engineering-team#650