fix 'NoneType' object is not subscriptable
Fix for https://l.codecov.dev/hqZ1Ci (Sentry issue)

The error itself is `TypeError: 'NoneType' object is not subscriptable`.
This was happening because if we don't have static analysis for a file
the lines relevant for that file might be `None`, and we we're not checking
this condition prior to starting to access the dict items.

The fix itself is very simple, but I wanted to add type hints and better
tests too, so they're in the change as well.