fix: match previous behavior around datapoints
As pointed out by @matt-codecov, there was a difference between the
old behaviour creates "outer list containing many single-entry inner
lists" whereas the new behavior was creating "outer list containing
a single large inner list".

This difference is subtle but it matters.
Each of the inner lists become a datapoint with all the labels in
the list.

Later on, when [deleting labels](https://github.com/codecov/shared/blob/main/shared/reports/editable.py#L52)
we delete a datapoint if _any_ of the labels to delete is in it.

This means that the old behavior made the datapoints stick around
for longer.
I should probably investigate if we ever merge datapoints... not sure

Anyway, matching old behavior.