don't save labels_index if no change
lables_index is used in the following scenarios:
1. merging reports during processing
2. label_analysis

In these cases we load it, use it, unload it.
Unloading includes saving it to storage again, and that is
how changes are persisted.

ONLY scenario 1 actually makes changes to it.
Now imagine that for some reason there's a label-analysis
request that takes a long time with the index open(unlikely)
and in that period an upload is processed (unlikely).
Then we would loose the updated index when label-analysis
re-writes it to storage.

To prevent that without a proper locking mechanism we can
just let label_analysis NOT save it back.
And because the merging scenario already runs inside a lock
we are fine :)