fix: check if labels before popping
A little bug was introduced in label-analysis as part of the label
encoding changes: to test if labels are encoded we pop one to see
if it's encoded or not.
In cases where the report has no labels this raises an exception.

These changes first see if there are labels before popping one.

Closes codecov/engineering-team#903