Create CleanLabelIndexTask (#205)
feat: CleanLabelsIndexTask

This task is designed to cleanup the labels_index of reports.
As they are carried forward, and tests are renamed, some indexes
will point to labels that are not used anymore. This will cause
the index to grow indefinetely and the lower indexes will eventually
not be used anymore.

THe task will scan a report and make sure all indexes are being used,
by modifying both the labels_index and the labels in the report.