fix: mark commits coming from forks in branch
Changes the branch name of a commit that is part of a PR from a fork.
This prevents branch collision issues where the coverage from a branch
in the target repo is overwritten by an upload coming from a fork.

The CLI has logic to add the branch, the API as well, but those are
ignored when uploading the commit info from provider.
Might as well build the logic here instead of trsuting CLI and API.