feat: create test results finisher task
This task will check if failures exist in the latest round
of test instances for this commit. If they do it will run
the test results notifier, else it will trigger the notify task
to run.

This task takes the same lock as the notify task so that they
don't try writing to the PR comment at the same time.

Signed-off-by: joseph-sentry <joseph.sawaya@sentry.io>