fix: rounding error in _apply_adjust_base_behavior
There's a bug where we compare the base adjusted coverage to the
head coverage. The issue occurs when the base adjusted coverage has more
digits after the decimal point than the head coverage.

The head coverage that we fetch in this function will always have 5
digits after the decimal, but the base adjusted coverage does not have
that same restriction.

The decimal library will not round the number with more digits after
the decimal point when doing a comparison.

Currently, in our code we are basically doing the following:

Decimal('90.001').quantize('0.00') <= Decimal('90.001')

which will evaluate to false.

If we are rounding the head coverage to 5 digits after the decimal point
we must do the same for the base adjusted coverage for it to be a fair
comparison.

Signed-off-by: joseph-sentry <joseph.sawaya@sentry.io>