Add test results finisher (#234)
* feat: create TestResultsNotifier class

This class creates the PR comment message based on the test instances
passed to it and posts/edits the comment on the PR.

* feat(notify): don't notify if failed test instances exist

The notify task will exit early now, if the latest test instances
relevant to the notify's commit contain failures

* feat: create test results finisher task

This task will check if failures exist in the latest round
of test instances for this commit. If they do it will run
the test results notifier, else it will trigger the notify task
to run.

This task takes the same lock as the notify task so that they
don't try writing to the PR comment at the same time.

* feat: run the test results finisher after the processors

The test result finisher runs after the test results processors
are done running. This is achieved using a celery chord.

* test: rename env to flags_hash in test_results_finisher

* fix: address feedback; inline True, remove notify var

* fix: incorporate insert_breaks feedback to simplify function

* fix: address feedback

this commit changes the int conversions of Outcome to
str conversions.

moves the db query for latest test instances for a given commit
to a separate function

and cleans up some unused vars

Signed-off-by: joseph-sentry <joseph.sawaya@sentry.io>