revamp sync_repos task
Adds a new flow for sync_repos task in which the caller
already knows what repos we should check. This happens
in INSTALLATION webhooks by GitHub, that include a list of affected repos.

The new flow makes use of github graphQL api via torngit to get info
directly on the repos affected.