fix: ignore insallation for owner some times
This is a problem with the 'sync repos' button in the UI.
It only affects getting the token for a OWNER (if it's for a repo it's correct).
Sadly this affects `sync_teams` and `sync_repos` tasks.

If the owner decides to sync we should use the owner's token,
even if there is an installation app for that owner.
The extra logic helps us do that.