chore: Change default RCB to adjust_base
These changes set the default Removed Code Behavior (RCB) to
`adjust_base` (instead of `fully_covered_patch`).

The change in the test is because the status actually passes
via adjust_base. I decided to preserve the outcome of the test,
and changed the rcb configured for the test.

context: codecov/engineering-team#1130