fix: Rewrite report in readable format in error
context: codecov/engineering-team#1112

Attempts to rewrite a report in readable format if the processing failed.
This is meant to aid in debugging errors - that likely come from the report
content - in case of processing failure.

Particularly the reports uploaded by the CLI are done so encoded and
compressed. If we rewrite in readable format users can download them in
the commits page of the UI and actually see what's inside.

Important details:
* This only applies if the processing failed with an exception.
* If we were to try processing the re-writen report the processing would
fail. (but we won't do that because processing failed with exception)
* If the error was an internal one the report won't be changed either.
Particularly for `SoftLimitTimeout` error it would be dangerous to
try and rewrite the report.