Bundle Analysis - update commit.state when finished processing report (#313)
we have this column in the DB commits.state, this gets flipped from pending to complete when a coverage report is done. When this field is flipped there's a DB that updates the branches table to the new branch->head.

This fixes the issue where the latest commit on the app is not showing the correct commit.