fix: check if PreProcessUpload is running
These changes add a verification for `PreProcessUpload` task already running
in another worker pod. This is to avoid a report being created twice for the
same commit.

Imagine the following scenario:
* You make 2 concurrent uploads (calling `create-report` twice for the same commit).
* Task A will get the lock and create a new report;
* Task A releases the lock and task B immediately gets it;
* Task A is commiting the transaction (outside the lock), at the same time task B passed [this point](https://github.com/codecov/worker/blob/e41ba11d68e37ffc7892e93954f68a7b0175cf93/services/report/__init__.py#L278) and is creating a new report;
* Task B creates a new report and saves it again;

I'm not sure if the report will have more sessions than it should,
but the database will show more `reports_upload` instances than it should

To avoid that we simply check if there's a task already running.
Because this task really only needs to run once, if one of them runs that's enough.