fix: propagate gh_installation name selection
After configuring multi-apps in staging we noticed the notificaitons were not coming
from the configured app. That is because each notifier uses its own `repo_service` instance.

These changes propagate the `gh_installation_name` selected by the Notify task
(the selection part is working properly) to the `NotificationService`.
The `NotificationService` in turn propagates the choice to the notifiers so that
when they get their own `repo_service` instances it will be from the correct installation.