chore: end the bulk insert experiment (#369)
Ends the bulk insert experiment. Uses only the bulk insert method to write tests,
for the reduced time.
This comes at the cost of using extra memory (aux structures), but the time gains
are worth it. Especially because right now the time to write to DB dominates the
total time of this task.