feat: pass fallbakc tokens to Github if available
depends on: https://github.com/codecov/shared/pull/186

Passes fallback tokens to GitHub instances if more than 1 app is available (in the case of dedicated apps in particular the default one is the fallback).
Also checks if apps are rate limited before selecting them.
There's a new exception now `NoConfiguredAppsAvailable` due to being rate limited.

I purposefully didn't catch this exception yet.
The idea in a subsequent PR is to do so in the notify task, for example,
so we can reschedule it.