feat: add error message to PR comment when test results parsing fails
The logic for doing this is as follows:
If there are no successful test results processing tasks this chord
then we check if there were previously test instances that were processed
for this commit.
If there are previous test instances on this commit, there is no error,
and we just exit this finisher without doing anything.
If there are no previous test instances on this commit, then we should
notify the user that there is an error. So we will begin by doing a
test results error comment which contains no info about coverage, then
we will queue up the notify task so it can create a coverage comment
that still contains this error's information.

If there are successful test results processing tasks, then we should
get rid of the error on the test results totals for this commit.

Signed-off-by: joseph-sentry <joseph.sawaya@sentry.io>