feat: retry notify if no apps are currently available
Catches `NoConfiguredAppsAvailable` exception in the notify task and
attempts to reschedule it at a later time.

The "later time" is in the next full hour, when the rate limits should reset.
Notice this only affects gh app installations, so tokenless should not fall unders this category.
(meaning we shouldn't have a terrible influx of tokenless notifs at the next hour).

Also updating the `get_utc_now` function because `datetime.utcnow` is deprecated.