fallback to assuming repo is covered
I suspect that we might have a little bug in some situations when syncing repos using integration.

Let's assume that suspicion is true and we might have repos not properly sinced in the installation.
These changes aim to not break the app in a specific scenario:
- Org installs Codecov App (not selecting all repos)
- Org admin logs in and sets up a new repo
- For some reason we fail to sync the repo and the installation list of repos. Maybe we missed a webhook? Would be the only possibility really... but maybe the admin logging in would actually make the repo appear in the DB, even missing the webhook.
- App is still broken cause the ORG never logged in, and the repo is not marked as "covered" by the installation

The changes here would fix that scenario by ignoring the repo when selecting the token to communicate with GH.
Ideally we fix the sync process and remove this fix, of course.

Notice that if the owner HAS a token or bot than we use that instead.