hotfix: sync_repos using integration when the repo services is known is acting strange (#405)
* hotfix: sync_repos using integration when the repo services is known is acting strange

* actually fix the sync issue

The error was a missing type cast on service_id which comes from the webhook as an int.

Also reactivating the previous logic (listing repos with installation) and potentially
updating the list of covered repos from that instead of relying exclusively on the webhook.
Considering that there are definetely some broken installations at this point it would give
the system ability to repair itself over time.