fix: get_repo_appropriate_bot_token returning None (#403)
get_repo_appropriate_bot_token is returning None in some cases which is
causing issues because it's expected to return a:
Tuple[Dict, Optional[Owner]].

I think the only way it can possibly return None is if get_public_token
returns None, which is possible if we don't enter the if statement in
get_public_token. So, if we don't enter it now, we now return a
RepositoryWithoutValidBotError exception.

Signed-off-by: joseph-sentry <joseph.sawaya@sentry.io>