fix: fix test_results_error comment logic
If there has been any success in processing for a given commit, it's
associated test result totals should have the error set to None,
otherwise it should be NO_SUCCESS. The condition for all_tests_passed
is: if the error is None and failed = 0 in the totals then all tests
passed. In the comment, if there's an error, show the error comment,
else if all tests passed show the all tests passed message. If neither
of these are true and the user HAS uploaded test results, then it means
that there must have been failures, or they somehow skipped all their
tests.

Signed-off-by: joseph-sentry <joseph.sawaya@sentry.io>