chore: extend UploadFlow with new ending
It was a suspition that we might be overcounting beginnings in the UploadFlow.
I believe one of the exit paths was not being covered in the `Upload` task.

These changes add a new ending to the UploadFlow. Namely, for the scenario where
an UploadTask finishes without enqueuing other tasks because there are no reports
waiting to be processed.

Also updating the `getutcnow` function that is deprecated and fixing some typos.

Changes to LockManager are probably from running 3.12 locally (new error appeared),
same thing for the test in `test_save_commit_measurements`