Remove unused import with ruff
In particular, I was interested in F401. This should clean up all unused imports in the repo.
We have to be careful with celery tasks because they don't get picked up if they are not "imported" anywhere.
I've added an annotation on the `__init__` files that import these "unused" tasks.