feat: add task gh_app_name override for more tasks (#436)
* feat: add task gh_app_name override for more tasks

This adds ability for more tasks to use the existing configuration options and
provide overrides for the GitHubAppInstallation name to be used for a given user.

Some of these usages have to be propagated through the task, as is the case with
ComputeComparison, for example.

:warning: These are not all instances, but most of them. Still working on others.

* feat: Override gh_app selected by ReportService

This is actually only relevant when we are creating a new report,
with carryforward. So should affect mostly `Upload` and `PreProcessUpload` tasks.

In any case I've added to some tasks that make intense use of reports.