refactor: unify interface to get auth info for torngit adapters
Because (again) of the recent multi-github app changes part of the logic that goes into
getting auth info for the torngit adapters was moved from `bots` service to the `repository`
and `owner` services (in their respective functions to get the provider adapter).

These changes unify that logic back into the `bots` service and put that as the interface
between `bots` and the other services through the `AdapterAuthInformation` type.
It should provide all the needed auth info for us to generate an appropriate torngit instance.

For what it's worth I beleive that the changes that happenend to the `owner` service related to
the token_refresh_callback were a bug that is fixed by these changes.
That is because the `owner.bot` is a different Owner that owns the token being used, so the callback
function needs to save the refreshed version back to `owner.bot`, not `owner`.