chore: remove deprecated installation_id check
We have been more than 15 days without seeing a log indicating that the `owner.integration_id`
was selected to communicate with github.

At this point the syncing of integrations is apparently OK and we have run the github backfill task
to create GithubAppInstallations to owners that didn't have those before.