refactor: Comment conditions
These are proposed changes to move much of the verifying logic of "can we send this comment?"
to a dedicated `NotificationCondition` class.

This is inspired by the authentication classes of rest framework (like we use in api), but
more explicit. My hope is that we can expand this for other notifiers as well, and hopefully
reuse some of the notify conditions accross notifiers.

But I'd like to test the waters (== team's appetite for these changes) before commiting
a bunch more time in refactoring more code.