feat: add function to get specific app's details
Adds `get_specific_github_app_details` function to bots service.
This allows us to get `GithubInstallationInfo` for the app we want.

It fails with a new exception if the app can't be found.
This is because this feature will be used to get the app that made a comment
or status check. If we don't use the same to edit said comment it won't work.

On top of that we need to start saving the ID of the apps we selected for Torngit adapter instances