feat: pin down commits to apps
extends NotificationResult to record the app that was used.
After notifications go through we save them to `CommitNotification`, including the app.
This is the "permanent pindown"
Previously this would only happen for pull requests, but not it _also_ saves a `CommitNotification` everytime
we have a record of an app_id being used (for a successful notification)

The `Notify` task after selecting an app will possibly pindown the
commit to that selected app, or refresh the permanent pindown in Redis (faster to access).
Notice then, that if a notifier is to use a specific app this will certainly
be in redis. Put there by the `NotifyTask`.

I've found some errors in the redis get/set though (through manual testing).
It turns out that the return value is a binary string, not a "regular" string.
Hence the changes to the "get" method.

Only notifiers that notify on a specific commit (status and checks) check for a specific app.
The comment notifier might use any app. But that's OK because an app _can_ edit
the comment of another app.