fix: bring back Owner.integration_id check
it aims to revert the important bits on https://github.com/codecov/worker/pull/445/commits/f58a5a07f14749e947ef02d66338f19c4043da28

That commit was released and we saw an immediate increase in errors of `RepoWithoutValidBot`.
We suspect that it was not correct to remove the deprecated path just yet.