fix: improve adjusted base comparison
ticket: https://github.com/codecov/internal-issues/issues/204

Let me preface this by saying that the latest comment in the ticket doesn't seem an accurate complain
cause looking in the logs the adjusted base coverage is 91.30752 and the head covereage is 91.24169.

That being said, it's still possible that for a decent sized code base the change is less than what we
display as a "change" (that is, the change is smaller than 0.01), but it's still a change (might be 0.0001)
Realistically no one would care.

So this change protects us by having a little wiggle room. The check would only fail if the difference is
more than 0.01 (so the numbers in the comment would look different).