refactor: Break up HeaderSectionWriter
As time when by the HeaderSectionWriter is now home for many warnings and details to the users,
and more are on the way.

This SectionWriter has a huge entrypoint function and it's hard to figure out the ordering of the
different lines. Also with the ability to hide project coverage it had more than 1 return value,
however there are messages that need to be written in both branches.

The changes in this commit attemp to unify the 2 branches of the HeaderSectionWriter, and breaks it up
further into HeaderPieces, with a function (or 2 😅) per HeaderPiece.

The objective is to make it easier to figure out the order of lines in the header.
AND to make it easier to change specific HeaderPieces and add more HeaderPieces.